Changed the way the config file is accessed and formatted
