+++ app/views.py
@@ -1,19 +1,17 @@
-from flask import send_from_directory, jsonify, json, render_template, redirect, request, session, flash, url_for
+from flask import render_template, redirect, request, session, url_for
 from app import app
-from app import db
 # Refactor later
-from app import models
 from app import spotifysso
-from app.API import spotify
-from app.tasks import update_user_tracks, get_last_n_minutes
-import os
-import config
+from app.API import spotify, influx
+from app.tasks import update_user_tracks
+
+
 # TODO: Remove this later
-from resources import query
 
 @app.route("/test")
 def indexer():
-    return render_template("dashboard.html",**locals())
+    return render_template("dashboard.html", **locals())
+
 
 @app.route("/index", methods=['GET', 'POST'])
 @app.route("/", methods=['GET', 'POST'])
@@ -21,15 +19,15 @@ def index():
     if "json_info" not in session:
         return render_template("login.html", **locals())
     else:
-        client = query.create_client('pse-ssh.diallom.com', 8086)
+        client = influx.create_client(app.config['INFLUX_HOST'], app.config['INFLUX_PORT'])
         userid = session['json_info']['id']
         access_token = spotify.get_access_token(session['json_info']['refresh_token'])
 
         recently_played = spotify.get_recently_played(access_token)
-        top_songs = query.get_top_songs(client, userid, 10, access_token)
-        top_genres = query.get_top_genres(client, userid, 10)
-        all_genres = query.get_genres(client, userid)
-        total_listening_time = query.total_time_spent(client, userid)
+        top_songs = influx.get_top_songs(client, userid, 10, access_token)
+        top_genres = influx.get_top_genres(client, userid, 10)
+        all_genres = influx.get_genres(client, userid)
+        total_listening_time = influx.total_time_spent(client, userid)
 
         # print(type(recently_played))
         # print(type(top_songs))
@@ -37,19 +35,19 @@ def index():
         # print(type(all_genres))
         # print(type(total_listening_time))
         # print(session['json_info'])
-        return render_template("index.html", **locals(), text=session['json_info']['display_name'], id=session['json_info']['id'])
-
+        return render_template("index.html", **locals(), text=session['json_info']['display_name'],
+                               id=session['json_info']['id'])
 
 
 @app.route("/index_js")
 def index_js():
-    client = query.create_client('pse-ssh.diallom.com', 8086)
+    client = influx.create_client(app.config['INFLUX_HOST'], app.config['INFLUX_PORT'])
     userid = session['json_info']['id']
     access_token = spotify.get_access_token(session['json_info']['refresh_token'])
 
-    top_songs = query.get_top_songs(client, userid, 10, access_token)
-    timestamps, duration = query.total_time_spent(client, userid)
-    top_genres = query.get_top_genres(client, userid, 10)
+    top_songs = influx.get_top_songs(client, userid, 10, access_token)
+    timestamps, duration = influx.total_time_spent(client, userid)
+    top_genres = influx.get_top_genres(client, userid, 10)
     duration = [time / 60000 for time in duration]
     songs, song_count = [list(x) for x in list(zip(*top_songs))]
     genres, genre_count = [list(x) for x in list(zip(*top_genres))]
@@ -61,7 +59,7 @@ def index_js():
 
 @app.route("/login")
 def login():
-    return spotifysso.authorize(callback="http://localhost:5000/callback")
+    return spotifysso.authorize(callback=f"http://{app.config['HOST']}:5000/callback")
 
     # return spotifysso.authorize(callback=url_for('authorized', _external=True, _scheme="https"))
